Support overrides for bin, lib, and include
Usually bin and include are the same, but for instance lib may want to be
lib64 on Red Hat distros.  Add BINDIR, LIBDIR, and INCLUDEDIR variables
that can be overridden on the command line.

This commit also tweaks the way the libraries are installed, dropping
execute privileges on libhoedown.a, and making sure libhoedown.so is
installed as a symlink.  (The install command doesn't preserve symlinks.)