Add special attributes for paragraphs.
This change also forces the @list specifier for list-level attributes.
Before, any string could have been used and it would have been
interpreted as a list-level special attribute.

Paragraphs that need special attributes must have the @paragraph
identifier in the beginning of their special attributes. This was used
because it is very common in text to have brackets, and would lead to
false positives otherwise.