Allow for header ID in case of special attribute.
Previously, the special attribute would override the header ID
extension. This change lets the two extensions play more nicely
together. If the special attribute string has an ID, it would override
the default header ID. If the special attribute string has no ID, then
the default header ID is used if the extension is turned on.