Add an option for an activation char for special attributes.
Sometimes, the current special attribute model can lead to accidental
attributes made with brackets. The behavior is currently not
configurable. This change allows the special attribute to require a
specified char in front of special attributes, to make it harder to make
it accidental.

For example, if the char was set to ':', then '# h {:.class}' would turn
into '<h1 class="class">h</h1>' but '# h {.class}' would not.