Fix offset computation when parsing list items.
When rendering list items as blocks, Hoedoewn's code assumed that the
item would always contain at least one newline. This assumption can
fail. For example, it does not hold when parsing footnotes in a file
that does not end in a trailing newline.

This change fixes the wrong assumption in the list rendering code. We
should also normalize text and add newlines *before* parsing footnotes;
however, that is a more complex change, and is the subject of hoextdown
issue 55 [1].

This bug was detected by the Google Autofuzz project.

[1]: https://github.com/kjdev/hoextdown/issues/55