Fix buffer-overread when checking for task lists.
The direct cause of the bug is that Hoedown does not check size
correctly when it tries to parse a loose task list. This leads to
problems when parsing tasks inside definition lists, where the
assumption that content is wrapped in <p> tags does not always hold.

We fix both the direct cause of the bug and add a check to ensure that
we do not aim to parse tasks in definition lists.

This bug was detected by the Google Autofuzz project.