Fix and improve handling of link attributes in Hoedown
This changes Hoedown to make link inline and not inline attributes
available through a side channel. This is done through a side channel
since changing the link callback API is not feasible (this is already a
pretty common pattern in the codebase).

By the way, my changes actually fixed a bug in Hoedown. The test case
that I changed was wrong before (it included "link-class" but should not
have). The bug before was char_link was mutating the non-inline
attribute and this was seen by subsequent links that used the same
reference link definition.