Re-write hoedown's rndr_attributes.
This started as an attempt to quickly fix a buffer overflow. Then I
realized that there are many cases, even common ones, that can break
this code. The present re-write still handles a couple of cases rather
badly (see test/Tests/extras/Attributes.text), but is more predictable
and safe than the original version.

For reference, here is the output of the previous code for some cases in
test/Tests/extras/Attributes.text:

    - valid and invalid cases {#id .cls foo=bar bar="baz" = !yippie =quux =}
      <li id="id" foo=bar b ar="baz" =  =quux = class="cls">valid and invalid cases</li>

    - single-letter class {class=a}
      <li class="">single-letter class</li>

    - two-letter class {class=ab}
      <li class="">two-letter class</li>

    - mixed with other attrs {w=xy #z}
      <li w=xy #z>mixed with other attrs</li>

    - we support extra spaces between attrs {   this=works   so="does this" .with-spaces }
      <li this=works   so="does this" .w>we support extra spaces between attrs</li>

The original buffer overflow was detected by the Google Autofuzz project.