Detect fenced code block starting in first line of list item
Prior to this change, the in_fenced flag was not set correctly for the first
line, and therefore inverted for every following line if the first line did
start with a code block. Since the start of a subsequent list item is
explicitly not detected while in fenced code, this essentially disabled the
has_next_oli detection, leading to HOEDOWN_LI_END terminating not only the
list item but the list as a whole.

Fixes https://github.com/hoedown/hoedown/issues/236.