Fix indentation of fenced code inside list item
In order to correctly recognize the closing fence of a fenced code block,
it is important to correctly determine the indentation of that code block.
If the code block starts in the first line of a list item, then there might
be more than one space after the list item, breaking the approach so far.

I have mentioned this problem in item 2 of https://github.com/kjdev/hoextdown/issues/59.