Ensure all required functions are included
Hoedown README claims that the library is pure C99, so it should be compiled
as such, not as the C90 implied by the -ansi flag. This will also avoid some
warnings about declaration after statement.

However in reality Hoextdown is using a number of functions that are not
part of the C99 standard: sntrncasecmp, snprintf, vsnprintf, isascii.
Setting the _DEFAULT_INCLUDES feature test macro makes these available from
glibc.