Restore attribute handling with empty block_id
Commit 08c3e4f48b2cf137f1848e789d6b5d21075b4e65 had been aiming to be a
semantic no-op in its attribute handling. But I had missed the fact that
with empty block_id any characters between '@' and ' ' would cause the match
to fail. This can be important e.g. when there are JavaDoc comments in code
as these are likely to start with "{@".

The sequence "{@ " can be used to introduce an extended comment inside a
code block. I'm not sure whether that's a bug or a feature, so I'm
documenting it in the test case for now, so that any change in behavior
there will at least be deliberate.